Extracted profiles related case classes to separate files
