+++ src/main/scala/com/softwaremill/realworld/profiles/Followers.scala
@@ -1,0 +1,3 @@
+package com.softwaremill.realworld.profiles
+
+case class Followers(userId: Int, followerId: Int)
