+++ src/main/scala/com/softwaremill/realworld/profiles/ProfileRow.scala
@@ -1,0 +1,3 @@
+package com.softwaremill.realworld.profiles
+
+case class ProfileRow(userId: Int, username: String, bio: String, image: String)
